MySurf with Chromium 39
It can run... but very unstable. Some basic features are still missing,
because they were lost in the move to Chromium 39. I am working on those
right now. Upgrading Chromium is not going to lose features or anything
in the future... it happened this time because we completely moved
projects.